requirements.txt
acme==1.1.0
amqp==5.2.0
async-timeout==4.0.3
black==23.11.0
blinker==1.7.0
Bootstrap-Flask==2.3.2
celery==5.3.6
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
openai-whisper==20231117
python-dotenv==1.0.0
redis==5.0.1
regex==2023.10.3
requests==2.31.0
setuptools-rust==1.8.1
tiktoken==0.5.1
tomli==2.0.1
torch==2.1.1
Werkzeug==3.0.1
WTForms==3.1.1
